requirements.txt
PyPDF2
openai>=1.0.0
python-dotenv
Flask
fpdf2
pandas
streamlit

# Offline MCQ Generation Dependencies
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
datasets>=2.12.0
sentencepiece>=0.2.0

# Professional MCQ Generation Dependencies (Optional - for best quality)
# Run 'python setup_professional_models.py' to install these
# transformers>=4.35.0
# torch>=2.0.0
# sentence-transformers>=2.2.0
# spacy>=3.7.0
# accelerate>=0.24.0
# bitsandbytes>=0.41.0